Fix: showing IDs in the dedicated server console.
New: Added password protection for private servers. Network protocol increased to 1104.

The password is asked in the "LoadingView".

For quick connect there are two ways to ask for password:

./netpanzer -c 1.2.3.4:3030 --password
./netpanzer -c netpanzer://1.2.3.4:3030/p

When --password or the url method with "/p" at the end is there, the LoadingView will ask for password before connecting to the server. The url way is usefull for the web server browser.

The info socket also sends new parameter if password is needed:

\password\{y,n}\

when is 'y' it needs password, when is 'n' it doesn't.

full example: gamename\netpanzer\protocol\1104\hostname\Player833\gameversion\svn-1293M\mapname\Cramped\mapcycle\Cramped, Tight Quarters, Two Villages\password\y\numplayers\0\maxplayers\8\empty\1\gamestyle\Objective\units_per_player\62\time\33\timelimit\50\fraglimit\500\objectivelimit\3\final\

In the in game server broswer, the servers has a small "red" key on the left if they need password, this is pretty ugly but is the best I could do.

The web serverbrowser should be modified to add to the netpanzer url "/p" at the end if the server requires password.

Server admins can set the password with:
/server gamepass <new_password>

and can query the current password with:
/server gamepass

in the config file is saved as:
game.gamepass = "thesecretpassword"

extra note: I know I didn't want to add any new function for 0.8.4 ... but Mathias asked...