- Fix sorting players by points correctly on ranking view and winner view - Fix crashing when selecting flags with right mouse button - Fix clossing the flag selection window without selecting flags - Fix correctly timing the flag change "since the last flag change" not when opening the window. - Fix keyboard not working correctly when mouse is over Help, Ranking and flag selection window - Added configuration to compile netpanzer for linux more portable - Added script to generate automatically a linux package file.
To create a package, go to the root of the netpazner code and create a file RELEASE_VERSION, the first line will be the release version:

# echo "0.8.4" > RELEASE_VERSION

Run the script to build the portable version:

# support/scripts/make-linux-portable.sh

When the script finished a new folder is created, name "releases", it will contain the file netpanzer-linux-<version>.tar.gz. Other players can uncompress this file and run the netpanzer executable

Example, after running the script, file is

netpanzer-linux-0.8.4.tar.gz

# tar zxf netpanzer-linux-0.8.4.tar.gz
# cd netpanzer
# ./netpanzer

The user is required to have SDL and SDL_Mixer installed.